third_party/blink/web_tests/fast/js/exception-line-number-expected.txt

This test makes sure the exception stack has line number information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "fast/js/script-tests/exception-line-number.js:6,fast/js/script-tests/exception-line-number.js:11"
PASS result is "fast/js/script-tests/exception-line-number.js:36"
PASS successfullyParsed is true

TEST COMPLETE
